Move compare_policies and hashable_policy functions into module_utils/ec2
